Decorator and permissions updates

+++ dojo/engagement/views.py
@@ -1378,6 +1378,9 @@ def edit_risk_acceptance(request, eid, raid):
 def view_edit_risk_acceptance(request, eid, raid, *, edit_mode=False):
     risk_acceptance = get_object_or_404(Risk_Acceptance, pk=raid)
     eng = get_object_or_404(Engagement, pk=eid)
+    # Ensure the risk acceptance belongs to the supplied engagement
+    if not Engagement.objects.filter(risk_acceptance=risk_acceptance, id=eid).exists():
+        raise PermissionDenied
 
     if edit_mode and not eng.product.enable_full_risk_acceptance:
         raise PermissionDenied
@@ -1539,6 +1542,9 @@ def expire_risk_acceptance(request, eid, raid):
     risk_acceptance = get_object_or_404(prefetch_for_expiration(Risk_Acceptance.objects.all()), pk=raid)
     # Validate the engagement ID exists before moving forward
     get_object_or_404(Engagement, pk=eid)
+    # Ensure the risk acceptance belongs to the supplied engagement
+    if not Engagement.objects.filter(risk_acceptance=risk_acceptance, id=eid).exists():
+        raise PermissionDenied
 
     ra_helper.expire_now(risk_acceptance)
 
@@ -1549,6 +1555,9 @@ def expire_risk_acceptance(request, eid, raid):
 def reinstate_risk_acceptance(request, eid, raid):
     risk_acceptance = get_object_or_404(prefetch_for_expiration(Risk_Acceptance.objects.all()), pk=raid)
     eng = get_object_or_404(Engagement, pk=eid)
+    # Ensure the risk acceptance belongs to the supplied engagement
+    if not Engagement.objects.filter(risk_acceptance=risk_acceptance, id=eid).exists():
+        raise PermissionDenied
 
     if not eng.product.enable_full_risk_acceptance:
         raise PermissionDenied
@@ -1562,6 +1571,9 @@ def reinstate_risk_acceptance(request, eid, raid):
 def delete_risk_acceptance(request, eid, raid):
     risk_acceptance = get_object_or_404(Risk_Acceptance, pk=raid)
     eng = get_object_or_404(Engagement, pk=eid)
+    # Ensure the risk acceptance belongs to the supplied engagement
+    if not Engagement.objects.filter(risk_acceptance=risk_acceptance, id=eid).exists():
+        raise PermissionDenied
 
     ra_helper.delete(eng, risk_acceptance)
